Module Js_of_ocaml_compiler.Config

module Flag : sig ... end

module Param : sig ... end

This module contains parameters that may be modified through command-line flags.

Parameters that are constant across a program run

These parameters should be set at most once at the beginning of the program.

val target : unit -> [ `JavaScript | `Wasm ]

val set_target : [ `JavaScript | `Wasm ] -> unit

type effects_backend = [

| `Disabled

| `Cps

| `Double_translation

| `Jspi

]

val effects : unit -> effects_backend

val set_effects_backend : effects_backend -> unit