Module Source_map.Mappings

type decoded = map list

type t

Represent the a list of mapping in its encoded form.

val empty : t

The empty mapping.

val is_empty : t -> bool

Test whether the mapping is empty.

val of_string_unsafe : string -> t

of_string_unsafe does not perform any validation of its argument, unlike decode_exn. It is guaranteed that of_string_unsafe and to_string are inverse functions. Time complexity O(1)

val decode_exn : t -> decoded

Parse the mappings.

val encode : decoded -> t

Encode the mappings.

val encode_with_offset : decoded -> Offset.t * t

Encode the mappings shifted by the returned offset so that the encoded mapping is more compact. This is useful to combining multiple mappings into an Index.t

val number_of_lines : t -> int

val first_line : t -> int

val to_string : t -> string

Returns the mappings as a string in the Source map v3 format. Time complexity O(1)