Module Js_of_ocaml_compiler.Structure

type graph

type t

val get_edges : graph -> Code.Addr.t -> Js_of_ocaml_compiler.Code.Addr.Set.t

val is_backward : t -> Code.Addr.t -> Code.Addr.t -> bool

val is_forward : t -> Code.Addr.t -> Code.Addr.t -> bool

val build_graph : Code.block Js_of_ocaml_compiler.Code.Addr.Map.t -> Code.Addr.t -> t

val dominator_tree : t -> graph

val is_merge_node : t -> Code.Addr.t -> bool

val is_loop_header : t -> Code.Addr.t -> bool

val sort_in_post_order : t -> Code.Addr.t list -> Code.Addr.t list

val blocks_in_reverse_post_order : t -> Code.Addr.t list

val get_nodes : t -> Js_of_ocaml_compiler.Code.Addr.Set.t

val norm : Code.program -> Code.program

norm p normalizes a program p to accommodate Structure.build_graph logic. In practice, it ensures that all loops have a predecessor block and allows to exit loops early.