Module Subst.Including_Binders

The operations of this module also substitute the variables names that appear on the left-hand-side of a Code.instr.Let, or as block parameters, or as closure parameters, or are bound by an exception handler.

val instr : (Code.Var.t -> Code.Var.t) -> Code.instr -> Code.instr

val instrs : (Code.Var.t -> Code.Var.t) -> Code.instr list -> Code.instr list

val block : (Code.Var.t -> Code.Var.t) -> Code.block -> Code.block

module And_Continuations : sig ... end