Module Js_of_ocaml_compiler__Base64

Base64 RFC4648 implementation.

Base64 is a group of similar binary-to-text encoding schemes that represent binary data in an ASCII string format by translating it into a radix-64 representation. It is specified in RFC 4648.

Release 3.9.1 - http://ocsigen.github.io/js_of_ocaml