Module Js_of_ocaml_compiler.Code

module Addr : sig ... end

module Var : sig ... end

type cont = Addr.t * Var.t list

type prim =

| Vectlength

| Array_get

| Extern of string

| Not

| IsInt

| Eq

| Neq

| Lt

| Le

| Ult

type array_or_not =

| Array

| NotArray

| Unknown

module Native_string : sig ... end

type constant =

| String of string

| NativeString of Native_string.t

| Float of Stdlib.Int64.t

| Float_array of Stdlib.Int64.t array

| Int of Targetint.t

| Int32 of Stdlib.Int32.t
(*
Only produced when compiling to WebAssembly.
*)

| Int64 of Stdlib.Int64.t

| NativeInt of Stdlib.Int32.t
(*
Only produced when compiling to WebAssembly.
*)

| Tuple of int * constant array * array_or_not

module Constant : sig ... end

type loc =

| No

| Before of Addr.t

| After of Addr.t

val noloc : loc

val location_of_pc : int -> loc

type prim_arg =

| Pv of Var.t

| Pc of constant

type special =

| Alias_prim of string

type mutability =

| Immutable

| Maybe_mutable

type field_type =

| Non_float

| Float

type expr =

| Apply of {

f : Var.t;

args : Var.t list;

exact : bool;

}

| Block of int * Var.t array * array_or_not * mutability

| Field of Var.t * int * field_type

| Closure of Var.t list * cont * Parse_info.t option

| Constant of constant

| Prim of prim * prim_arg list

| Special of special

type instr =

| Let of Var.t * expr

| Assign of Var.t * Var.t

| Set_field of Var.t * int * field_type * Var.t

| Offset_ref of Var.t * int

| Array_set of Var.t * Var.t * Var.t

| Event of Parse_info.t

type last =

| Return of Var.t

| Raise of Var.t * [ `Normal | `Notrace | `Reraise ]

| Stop

| Branch of cont

| Cond of Var.t * cont * cont

| Switch of Var.t * cont array

| Pushtrap of cont * Var.t * cont

| Poptrap of cont

type block = {

params : Var.t list;

body : instr list;

branch : last;

}

type program = {

start : Addr.t;

blocks : block Addr.Map.t;

free_pc : Addr.t;

}

module Print : sig ... end

type 'c fold_blocs = block Addr.Map.t -> Addr.t -> (Addr.t -> 'c -> 'c) -> 'c -> 'c

type fold_blocs_poly = {

fold : 'a. 'a fold_blocs;

}

val fold_closures : program -> (Var.t option -> Var.t list -> cont -> Parse_info.t option -> 'd -> 'd) -> 'd -> 'd

fold_closures p f init folds f over all closures in the program p, starting from the initial value init. For each closure, f is called with the following arguments: the closure name (enclosed in Stdlib.Some), its parameter list, the address and parameter instantiation of its first block, the optional closure location and the current accumulator. In addition, f is called on the initial block p.start, with None as the closure name. All closures in all blocks of p are included in the fold, not only the ones reachable from p.start.

val fold_closures_innermost_first : program -> (Var.t option -> Var.t list -> cont -> Parse_info.t option -> 'd -> 'd) -> 'd -> 'd

Similar to fold_closures, but applies the fold function to the innermost closures first. Unlike with fold_closures, only the closures reachable from p.start are considered.

val fold_closures_outermost_first : program -> (Var.t option -> Var.t list -> cont -> Parse_info.t option -> 'd -> 'd) -> 'd -> 'd

Similar to fold_closures, but applies the fold function to the outermost closures first. Unlike with fold_closures, only the closures reachable from p.start are considered.

val fold_children : 'c fold_blocs

val fold_children_skip_try_body : 'c fold_blocs

val poptraps : block Addr.Map.t -> Addr.t -> Addr.Set.t

val traverse : fold_blocs_poly -> (Addr.t -> 'c -> 'c) -> Addr.t -> block Addr.Map.t -> 'c -> 'c

val preorder_traverse : fold_blocs_poly -> (Addr.t -> 'c -> 'c) -> Addr.t -> block Addr.Map.t -> 'c -> 'c

val last_instr : instr list -> instr option

Last instruction of a block body, ignoring events

val used_blocks : program -> Stdlib.BitSet.t

val prepend : program -> instr list -> program

val empty : program

val compact : program -> program

val is_empty : program -> bool

val equal : program -> program -> bool

val print_diff : program -> program -> unit

val check_updates : name:string -> program -> program -> updates:int -> unit

val invariant : program -> unit

val cont_equal : cont -> cont -> bool

val cont_compare : cont -> cont -> int