Module Source_map.Mappings

type t

val empty : t

Represents the empty mapping.

val of_string : string -> t

By default, mappings are left uninterpreted, since many operations can be performed efficiently directly on the encoded form. Therefore this function is mostly a no-op and very cheap. It does not perform any validation of its argument, unlike decode. It is guaranteed that of_string and to_string are inverse functions.

val decode : t -> map list

Parse the mappings.

val encode : map list -> t

val to_string : t -> string

Returns the mappings as a string in the Source map v3 format. This function is mostly a no-op and is very cheap.